+++ src/types/event.ts
@@ -1,12 +1,12 @@
 import Theme from '@/styles/Theme';
 
+import { FormType } from './form';
+
 type EventStatus = 'CONFIRMED' | 'PENDING' | 'CANCEL_REQUESTED' | 'CANCELLED';
 
 type EventTagKey = 'publicEvent' | 'clubOnlyEvent' | EventStatus;
 
 type EventType = 'performance' | 'promotion' | 'recruitment' | 'clubSchedule';
-
-type FormType = 'RADIO' | 'TEXT' | 'SELECT' | 'NUMBER';
 
 type ProfileEventType = 'applied' | 'bookmark';
 
@@ -57,13 +57,6 @@ interface GetAllEventsResponse {
   };
 }
 
-interface FormOption {
-  id: string;
-  title: string;
-  type: FormType;
-  visible: boolean;
-}
-
 interface getEventDetailResponse {
   title: string;
   posterImageUrl: string;
@@ -98,8 +91,6 @@ export {
   GetAllEventsRequest,
   GetAllEventsResponse,
   getEventDetailResponse,
-  FormType,
-  FormOption,
   Event,
   EventTags,
   EventTagKey,
