+++ src/types/form.ts
@@ -1,0 +1,10 @@
+type FormType = 'RADIO' | 'TEXT' | 'SELECT' | 'NUMBER';
+
+interface FormOption {
+  id: string;
+  title: string;
+  type: FormType;
+  visible: boolean;
+}
+
+export { FormType, FormOption };
